Modulo de vacaciones completo sin errores.

+++ View/crearDepartamento.php
@@ -4,6 +4,7 @@ include_once('../Controller/insertar.php');
 #include_once('Controller/eliminar.php');
 
 $Departamentos = getDepartamentos();
+$NombreE = getNombreE();
 
 $_SESSION['titulo'] = 'Departamento';
 include('../include/header.php');
@@ -26,9 +27,12 @@ $_SESSION["inicio"] = "../inicio.php";
 $_SESSION["departamento"] = "crearDepartamento.php";
 $_SESSION["encargado"] = "crearEncargado.php";
 $_SESSION["empleado"] = "crear.php";
+$_SESSION['Vacaciones'] = "vacaciones.php";
+
 $_SESSION["empleadoA"] = "";
 $_SESSION["encargadoA"] = "";
 $_SESSION["departamentoA"] = "active";
+$_SESSION['VacacionesA'] = "";
 include('../include/nav.php');
 ?>
 
@@ -45,7 +49,13 @@ include('../include/nav.php');
                 <div class="modal-body">
                     <div class="row mb-5">
                         <div class="col-4">
-                            <input type="number" class="form-control" placeholder="ID del Encargado" name="encargado_id" required>
+                            <!--<input type="number" class="form-control" placeholder="ID del Encargado" name="encargado_id" required> -->
+                            <select class=" form-select" name="encargado_id">
+                                <option disabled selected>...</option>
+                                <?php while ($fila = $NombreE->fetch_object()) { ?>
+                                    <option value="<?php echo $fila->ID; ?>"><?php echo $fila->Encargado; ?></option>
+                                <?php } ?>
+                            </select>
                         </div>
                         <div class="col-4">
                             <input type="text" class="form-control" placeholder="Nombre" name="nombre" required>
@@ -71,30 +81,30 @@ include('../include/nav.php');
     <!-- Button del modal -->
     <a href="View/crearDepartamento.php" class="btn btn-success" data-bs-toggle="modal" data-bs-target="#exampleModal"><i class="fas fa-plus-circle"></i> Crear Departamento</a>
     <div class="table-responsive">
-        <table id="tblDepart" class="table table-hover table-striped">
+        <table id="tblDepart" class="table table-hover table-striped table-fixed">
             <thead>
                 <tr>
-                    <th>ID </th>
-                    <th>Encargado </th>
-                    <th>Departamento</th>
-                    <th>Descripcion</th>
-                    <th>Opciones</th>
+                    <th style="width: 250px;">Encargado </th>
+                    <th style="width: 300px;">Departamento</th>
+                    <th style="width: 580px;">Descripcion</th>
+                    <th style="width: 150px; text-align:center">Opciones</th>
                 </tr>
             </thead>
-            <?php while ($fila = $Departamentos->fetch_object()) { ?>
-                <tr>
-                    <td><?php echo $fila->ID; ?></td>
-                    <td><?php echo $fila->Encargado; ?></td>
-                    <td style="width: 200px;"><?php echo $fila->Departamento; ?></td>
-                    <td><?php echo $fila->Descripcion; ?></td>
-                    <td style="width: 150px;">
-                        <button class="btn btn-outline-primary btn-sn" title="Editar registro"><i class="fas fa-user-edit"></i></button>
-                        <a href="../Controller/eliminar.php?Entidad=departamento&ID=<?php echo $fila->ID; ?>" class="btn btn-outline-danger btn-sn"
-                        title="Eliminar registro" onclick="return confirm('Eliminar un encargado puede provocar que se eliminen registros dependientes.')"><i class="fas fa-trash-alt"></i></a>
-                    </td>
-                </tr>
-            <?php } ?>
+            <tbody>
+                <?php while ($fila = $Departamentos->fetch_object()) { ?>
+                    <tr>
+                        <td style="width: 250px;"><?php echo $fila->Encargado; ?></td>
+                        <td style="width: 300px;"> <?php echo $fila->Departamento; ?></td>
+                        <td style="width: 580px;"><?php echo $fila->Descripcion; ?></td>
+                        <td style="width: 150px; text-align:center">
+                            <button class="btn btn-outline-primary btn-sn" title="Editar registro"><i class="fas fa-user-edit"></i></button>
+                            <a href="../Controller/eliminar.php?Entidad=departamento&ID=<?php echo $fila->ID; ?>" class="btn btn-outline-danger btn-sn" title="Eliminar registro" onclick="return confirm('Eliminar un encargado puede provocar que se eliminen registros dependientes.')"><i class="fas fa-trash-alt"></i></a>
+                        </td>
+                    </tr>
+                <?php } ?>
+            </tbody>
         </table>
+    </div>
 </main>
